TGEN: Added implementation of OMPL framework
Sub8StateSpace is an extension of OMPL's CompoundStateSpace.
The sub's state space is defined as a 12D vector. Currently,
the time derivative of the euler angles is used to be
the representation of the angular velocities,
and a unit quaternion is used for the orientation.

Planning will be carried out in the sub's control space;
accordingly, Sub8SpaceInformation is an extension of OMPL's
control::SpaceInformation.

StateValidityChecker allows the client to customize
collision checking, clearance from nearest invalid states,
etc.

Most unit tests are currently failing, since tests have
been written for unimplemented functions for Sub specific
behavior.

catkin_make was verified to build successfully.