TGEN: Integrated MotionPlan and Waypoint msgs
Implemented the TGenNode class, which provides
a callback for a motion planner service. The
mission planner can request a trajectory when
providing a start and goal Waypoint, and receive
a trajectory in return.

More clang formatting in the test directory.

Added a method in TGenManager that converts a
Waypoint to State.

Tests that create a Sub8SpaceInformation object fail
bc the methods for setting bounds on the subspaces
haven't been implemented yet.

Catkin_make verified to build successfully.

Note: catkin_make was failing to find the sub8_msgs
package when I tried to include a generated header
for the MotionPlan service. Removing the
LIBRARIES sub8_msgs line in catkin_package(...) in the
CMakeLists.txt in the sub8_msgs package fixed the problem.
Other devs should verify that they can use sub8_msgs stil.