TGEN: postPropagate, regularize rate change of orientation
Added a regularization term when computing the
rate change of orientation. If the current orientation
is not a unit quaternion, the added term will
attemptto drive it back to having a norm of 1.

Implemented the simple postPropagate method,
which re-normalizes the orientation. Both
this and the regularization should ensure
we keep our orientation-game on 100