TGEN: Implemented a more sophisticated distance function
In preparation for end-to-end testing of the TGEN,
I updated the distance function to a more sophisticated
metric that uses a weighted euclidean distance between
two states. See the implementation for more details.

Also did some minor refactoring to the Sub8StateSpace.

Updated unit and integration tests to reflect
all changes.