GREAT MERGE: deprecate Float64Stamped.msg
Replaced with RangeStamped. DVL related cold that previously used
Float64Stamped should be transitioned to use RangeStamped instead.

Other depth driver related code should be changed to use
DepthStamped.

The only usage of Float64Stamped that is not DVL or depth or depth
driver related is in the start gate mission and should be removed
(@RustyBamboo).